requirements.txt
Django==4.1
djangorestframework
pyyaml
requests
django-cors-headers
